fix comments, naming
