+++ csrc/host_ir/executor.h
@@ -14,15 +14,13 @@
 #ifdef NVFUSER_HOST_IR_JIT
 #include <host_ir/jit.h>
 #endif
+#include <c10/cuda/CUDAStream.h>
 #include <multidevice/communicator.h>
 #include <multidevice/ipc_handle.h>
 #include <runtime/executor.h>
 #include <runtime/executor_abstract.h>
 #include <runtime/executor_params.h>
 #include <runtime/fusion_executor_cache.h>
-#include <utils.h>
-#include <c10/cuda/CUDAStream.h>
-#include <host_ir/params.h>
 
 namespace nvfuser {
 
@@ -53,6 +51,23 @@ class HostIrExecutor : public ExecutorAbstract {
 };
 
 namespace hir {
+
+// Set of parameters that control the behavior of HostIrEvaluator
+struct HostIrEvaluatorParams {
+  // Experimental: whether to use FusionExecutorCache rather than
+  // KernelExecutor.
+  bool use_fusion_executor_cache = false;
+  // Experimental: whether to apply auto-scheduling in FusionExecutorCache if
+  // use_fusion_executor_cache=true. WAR: temporary hack mainly use for
+  // development
+  bool skip_auto_scheduling = false;
+  // Experimental: whether to cache fusion executor. WAR: avoid recompilation
+  // but implicitely assumes that the input shape don't change over iterations
+  bool cache_fusion_executor = false;
+  // number of additional cuda streams to use at runtime for comm+compute
+  // pipelining
+  int64_t number_of_streams = 4;
+};
 
 // A HostIrEvaluator evaluates a host programs represented through a
 // HostIrContainer It is instantiated with the desired HostIrContainer, and runs
